fix some imports after moving region
