+++ hydromt/region/__init__.py
@@ -1,5 +1,5 @@
 """Model for dealing with region specifcation and manipulation."""
-from hydromt.region import Region
+from hydromt.region.region import Region
 
 __all__ = [
     "Region",
