+++ tests/region/test_model_region.py
@@ -20,7 +20,7 @@ def test_bbox_region():
 def test_region_unknown_key_errors():
     region = {"region": [0.0, -1.0]}
     with pytest.raises(ValueError, match=r"Unknown region kind.*"):
-        _ = Region(region, data_catalog=None)
+        _ = Region(region)
 
 
 @pytest.mark.skip(reason="model region spec not yet implemented")
